Move all buffer related classes into the buffer package.
